requirements.txt
streamlit==1.28.1
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
gitpython==3.1.40
